# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubscribeRequest {
  /** @maxLength 254 */
  email: string;
}

export interface SubscribeResult {
  ok: boolean;
  message: string;
}

export interface UnsubscribeResult {
  ok: boolean;
  message: string;
}

export interface ApiError {
  error: string;
}

export type UnsubscribeParams = {
  token: string;
};