# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: subscribers
    description: Email subscriber operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /subscribe:
    post:
      operationId: subscribe
      tags: [subscribers]
      summary: Subscribe to the newsletter
      description: Adds an email to the subscriber list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeRequest"
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResult"
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /unsubscribe:
    get:
      operationId: unsubscribe
      tags: [subscribers]
      summary: Unsubscribe via token
      description: Removes a subscriber given an unsubscribe token
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Unsubscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsubscribeResult"
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SubscribeRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
      required:
        - email
    SubscribeResult:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
      required:
        - ok
        - message
    UnsubscribeResult:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
      required:
        - ok
        - message
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error